requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 2.2.2
Flask_Bcrypt == 1.0.1
Flask_Login == 0.6.2
Flask_WTF == 1.0.1
Jinja2 == 3.1.2
WTForms == 3.0.1
app == 0.0.1
flask-sqlalchemy-meiqia == 2016.8.1
flask-sqlalchemy-tw == 3.0-dev-20160331
netoprmgr == 1.3.5
python_decouple == 3.6
